+++ web-ui/src/components/common/BlobDetailsDialog/BlobDetailsDialogUi.tsx
@@ -20,6 +20,7 @@ import { BlobStatsDto } from '../../../../generated';
 import { JSX } from 'react/jsx-runtime';
 import { useTranslation } from 'react-i18next';
 import { StateIcon } from './StateIcon';
+import { TraitIcon } from './TraitIcon';
 
 interface BlobDetailsDialogUiProps {
   open: boolean;
@@ -44,9 +45,16 @@ export const BlobDetailsDialogUi = ({ open, onClose, blob, blobIcon }: BlobDetai
         <Box display="flex" flexDirection="column" alignItems="center" gap={1}>
           <Box display="flex" flexDirection="column" alignItems="center" width="100%">
             <Box display="flex" width="100%">
-              {blob.states.map((state) => (
-                <StateIcon key={state.type} state={state} />
-              ))}
+              <Box display="flex" width="100%">
+                {blob.states.map((state) => (
+                  <StateIcon key={state.type} state={state} />
+                ))}
+              </Box>
+              <Box display="flex" width="100%" justifyContent="flex-end">
+                {blob.traits.map((trait) => (
+                  <TraitIcon key={trait} trait={trait} />
+                ))}
+              </Box>
             </Box>
             {blobIcon}
           </Box>