+++ web-ui/src/components/common/blob-visuals/blob-parts/eyes/EyeStars.tsx
@@ -1,0 +1,24 @@
+export const EyeStars = () => {
+  return (
+    <>
+      <path
+        fill="#ffcc00"
+        stroke="#000000"
+        strokeWidth={0}
+        strokeLinecap="butt"
+        strokeOpacity={1}
+        d="M 85,26 c 0,9 -4,9 -4,9 c 4,0 4,9 4,9 c 0,-9 4,-9 4,-9 c -4,0 -4,-9 -4,-9 Z"
+        id="start2"
+      />
+      <path
+        fill="#ffcc00"
+        stroke="#000000"
+        strokeWidth={0}
+        strokeLinecap="butt"
+        strokeOpacity={1}
+        d="M 45,26 c 0,9 -4,9 -4,9 c 4,0 4,9 4,9 c 0,-9 4,-9 4,-9 c -4,0 -4,-9 -4,-9 Z"
+        id="start2-6"
+      />
+    </>
+  );
+};