+++ domain/dtos/blob_stats_dto.py
@@ -2,7 +2,9 @@ import enum
 from pydantic import BaseModel
 
 from domain.dtos.parent_dto import ParentDto
+from domain.dtos.state_dto import StateDto
 from domain.enums.activity_type import ActivityType
+from domain.enums.trait_type import TraitTypeDto
 
 
 class IntegrityState(enum.Enum):
@@ -34,3 +36,5 @@ class BlobStatsDto(BaseModel):
     integrity_color: str | None = None
     current_activity: ActivityType | None = None
     current_standings_position: int | None = None
+    states: list[StateDto]
+    traits: list[TraitTypeDto]