+++ domain/dtos/trait_type_dto.py
@@ -1,0 +1,4 @@
+from data.model.trait_type import TraitType
+
+
+TraitTypeDto = TraitType