+++ web-ui/src/components/common/blob-visuals/blob-parts/BlobBase.tsx
@@ -8,9 +8,10 @@ type BlobBaseProps = {
   hasCrown: boolean;
   eyeSlot: JSX.Element;
   toolSlot?: JSX.Element;
+  bruiseSlot?: JSX.Element;
 };
 
-export const BlobBase = ({ size, color, eyeSlot, doSquash, hasCrown, toolSlot }: BlobBaseProps) => {
+export const BlobBase = ({ size, color, eyeSlot, doSquash, hasCrown, toolSlot, bruiseSlot }: BlobBaseProps) => {
   return (
     <svg width={size} height={size} viewBox="0 0 130 95" version="1.1" id="svg5" xmlns="http://www.w3.org/2000/svg">
       <g id="layer1">
@@ -33,6 +34,7 @@ export const BlobBase = ({ size, color, eyeSlot, doSquash, hasCrown, toolSlot }:
               d="m 129.5089,59.689952 c 0,27.20034 -28.7716,34.818936 -64.263171,34.818937 -35.491565,0 -64.7546185,-7.618596 -64.7546185,-34.818937 C 0.49111025,32.48961 29.754163,0.4911092 65.245729,0.49110935 100.7373,0.49111061 129.5089,32.489611 129.5089,59.689952 Z"
             />
             {eyeSlot}
+            {bruiseSlot}
           </g>
 
           {/* Crown is a sibling of the scaled group so it won't be scaled.