+++ web-ui/src/components/common/blob-visuals/blob-parts/eyes/GloomyEyebrows.tsx
@@ -1,0 +1,26 @@
+export const GloomyEyebrows = () => {
+  return (
+    <>
+      <path
+        fill="none"
+        stroke="#000000"
+        stroke-width="1"
+        stroke-linecap="butt"
+        stroke-linejoin="miter"
+        stroke-opacity="1"
+        d="m 45,22 -7,12"
+        id="eyebrow1"
+      />
+      <path
+        fill="none"
+        stroke="#000000"
+        stroke-width="1"
+        stroke-linecap="butt"
+        stroke-linejoin="miter"
+        stroke-opacity="1"
+        d="m 85,22 7,12"
+        id="eyebrow2"
+      />
+    </>
+  );
+};